backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
azure-ai-agents==1.1.0
azure-ai-inference==1.0.0b9
azure-ai-projects==1.0.0b12
azure-common==1.1.28
azure-core==1.35.0
azure-core-tracing-opentelemetry==1.0.0b12
azure-cosmos==4.14.0b2
azure-identity==1.24.0
azure-search-documents==11.5.3
azure-storage-blob==12.26.0
bcrypt==4.0.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
cryptography==45.0.6
dataclasses-json==0.6.7
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
eval_type_backport==0.2.2
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
filelock==3.19.1
frozenlist==1.7.0
greenlet==3.2.4
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jwt==1.4.0
langchain==0.3.27
langchain-azure-ai==0.1.5
langchain-community==0.3.29
langchain-core==0.3.75
langchain-openai==0.3.30
langchain-pinecone==0.2.11
langchain-tests==0.3.20
langchain-text-splitters==0.3.9
langchain-together==0.3.1
langgraph==0.6.6
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.5
langsmith==0.4.14
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
msal==1.33.0
msal-extensions==1.3.1
multidict==6.6.4
mypy_extensions==1.1.0
numpy==2.3.2
openai==1.100.2
opentelemetry-api==1.36.0
orjson==3.11.2
ormsgpack==1.10.0
packaging==24.2
pandas==2.3.2
passlib==1.7.4
pillow==11.3.0
pinecone==7.3.0
pinecone-client==6.0.0
pinecone-plugin-assistant==1.7.0
pinecone-plugin-interface==0.0.7
pluggy==1.6.0
postgrest==1.1.1
propcache==0.3.2
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pymssql==2.3.7
PyMuPDF==1.26.4
pyodbc==5.2.0
pytest==8.4.1
pytest-asyncio==0.26.0
pytest-benchmark==5.1.0
pytest-codspeed==4.0.0
pytest-recording==0.13.4
pytest-socket==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
realtime==2.7.0
regex==2025.7.34
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
rich-toolkit==0.15.0
rignore==0.6.4
rsa==4.9.1
sentry-sdk==2.35.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.2
storage3==0.12.1
StrEnum==0.4.15
supabase==2.18.1
supabase_auth==2.12.3
supabase_functions==0.10.1
syrupy==4.9.1
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.11.0
together==1.5.25
tqdm==4.67.1
typer==0.15.4
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.35.0
vcrpy==7.0.0
watchfiles==1.1.0
websockets==15.0.1
wrapt==1.17.3
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.24.0
